Inteview\Invariants\Tree: extracted cascading parent answer correspondance invariant KP-7775

+++ src/Core/SharedKernels/DataCollection/DataCollection/Aggregates/IQuestionnaire.cs
@@ -70,7 +70,7 @@ namespace WB.Core.SharedKernels.DataCollection.Aggregates
 
         string GetAnswerOptionTitle(Guid questionId, decimal answerOptionValue);
 
-        decimal GetCascadingParentValue(Guid questionId, decimal answerOptionValue);
+        int GetCascadingParentValue(Guid questionId, decimal answerOptionValue);
 
         int? GetMaxSelectedAnswerOptions(Guid questionId);
